requirements.txt
flask 
ariadne 
flask-sqlalchemy 
flask-cors
python-dotenv
.env
.vscode